requirements.txt
flask
waitress
requests
python-dotenv
